Solves the missing cursor issue
pywal would set the st cursor the same color as the background of the terminal making it difficult to firstly identify where the cursor really is and secondly if it were on a text, the letter would appear as a blank space than the color of the text changing signifying that the cursor has engulfed the letter. really cool patch thank you suckless devs for solving this issue which I couldnt find a workaround for because pywal/xresources state my cursor color to be some white/grey hex color code when it clearly is not.